move gRPC-specific bootstrap code into its own module
