+++ src/core/ext/xds/xds_common_types.cc
@@ -38,7 +38,7 @@
 #include "xds/type/v3/typed_struct.upb.h"
 
 #include "src/core/ext/xds/upb_utils.h"
-#include "src/core/ext/xds/xds_bootstrap.h"
+#include "src/core/ext/xds/xds_bootstrap_grpc.h"
 #include "src/core/ext/xds/xds_client.h"
 
 namespace grpc_core {
@@ -125,7 +125,9 @@ CertificateProviderInstanceParse(
           UpbStringToStdString(
               envoy_extensions_transport_sockets_tls_v3_CommonTlsContext_CertificateProviderInstance_certificate_name(
                   certificate_provider_instance_proto))};
-  if (!context.client->bootstrap().certificate_provider_plugin_map()->HasPlugin(
+  const auto& bootstrap =
+      static_cast<const GrpcXdsBootstrap&>(context.client->bootstrap());
+  if (!bootstrap.certificate_provider_plugin_map()->HasPlugin(
           certificate_provider_plugin_instance.instance_name)) {
     return absl::InvalidArgumentError(
         absl::StrCat("Unrecognized certificate provider instance name: ",
@@ -147,7 +149,9 @@ CertificateProviderPluginInstanceParse(
           UpbStringToStdString(
               envoy_extensions_transport_sockets_tls_v3_CertificateProviderPluginInstance_certificate_name(
                   certificate_provider_plugin_instance_proto))};
-  if (!context.client->bootstrap().certificate_provider_plugin_map()->HasPlugin(
+  const auto& bootstrap =
+      static_cast<const GrpcXdsBootstrap&>(context.client->bootstrap());
+  if (!bootstrap.certificate_provider_plugin_map()->HasPlugin(
           certificate_provider_plugin_instance.instance_name)) {
     return absl::InvalidArgumentError(
         absl::StrCat("Unrecognized certificate provider instance name: ",
